Grahamc/feature/linuxkit builder

+++ pkgs/build-support/vm/linuxkit-builder/default.nix
@@ -16,9 +16,11 @@
 , hyperkit
 , vpnkit
 , linuxkit
+, buildEnv
 , writeScript
 , writeScriptBin
 , writeText
+, writeTextFile
 , forceSystem
 , vmTools
 , makeInitrd
@@ -26,12 +28,13 @@
 , linuxkitKernel ? import ./kernel.nix { pkgs = forceSystem "x86_64-linux" "x86_64"; }
 , storeDir ? builtins.storeDir
 
-, authorizedKeys ? ./default-key.pub
-, privateKey ? ./default-key
 , hostPort ? "24083"
 }:
 
 let
+  writeScriptDir = name: text: writeTextFile {inherit name text; executable = true; destination = "/${name}"; };
+  writeRunitForegroundService = name: run: writeTextFile {inherit name; text = run; executable = true; destination = "/${name}/run"; };
+
   pkgsLinux = forceSystem "x86_64-linux" "x86_64";
   vmToolsLinux = vmTools.override { kernel = linuxkitKernel; pkgs = pkgsLinux; };
   containerIp = "192.168.65.2";
@@ -73,10 +76,10 @@ let
 
     mkdir /fs
 
-    mount /dev/${hd} /fs 2>/dev/null || {
+    if ! mount /dev/${hd} /fs 2>/dev/null; then
       ${pkgsLinux.e2fsprogs}/bin/mkfs.ext4 -q /dev/${hd}
       mount /dev/${hd} /fs
-    } || true
+    fi
 
     mkdir -p /fs/dev
     mount -o bind /dev /fs/dev
@@ -111,8 +114,117 @@ let
     PasswordAuthentication no
     ChallengeResponseAuthentication no
   '';
-  stage2Init = writeScript "vm-run-stage2" ''
-    #! ${pkgsLinux.bash}/bin/bash
+stage2Init = let
+    script_modprobe = writeScript "modeprobe" ''
+      #! /bin/sh
+      export MODULE_DIR=${pkgsLinux.linux}/lib/modules/
+      exec ${pkgsLinux.kmod}/bin/modprobe "$@"
+    '';
+
+    file_passwd = writeText "passwd" ''
+      root:x:0:0:System administrator:/root:${pkgsLinux.bash}/bin/bash
+      sshd:x:1:65534:SSH privilege separation user:/var/empty:${pkgsLinux.shadow}/bin/nologin
+      nixbld1:x:30001:30000:Nix build user 1:/var/empty:${pkgsLinux.shadow}/bin/nologin
+    '';
+
+    file_group = writeText "group" ''
+      nixbld:x:30000:nixbld1
+      root:x:0:root
+    '';
+
+    file_bashrc = writeText "bashrc" ''
+      export PATH="${vmToolsLinux.initrdUtils}/bin:${pkgsLinux.nix}/bin"
+      export NIX_SSL_CERT_FILE='${pkgsLinux.cacert}/etc/ssl/certs/ca-bundle.crt'
+    '';
+
+    script_poweroff = writeText "poweroff" ''
+      #!/bin/sh
+      exec ${pkgsLinux.busybox}/bin/poweroff
+    '';
+
+    script_poweroff_f = writeText "poweroff" ''
+      #!/bin/sh
+      exec ${pkgsLinux.busybox}/bin/poweroff -f
+    '';
+
+    file_instructions = writeText "instructions" ''
+      ======================================================================
+      Remote builder has started.
+
+      If this is a fresh VM you need to run the following on the host:
+          ~/.nixpkgs/linuxkit-builder/finish-setup.sh
+
+
+      Exit this VM by running:
+          kill $(cat ~/.nixpkgs/linuxkit-builder/nix-state/hyperkit.pid)
+      ======================================================================
+    '';
+
+    runit_targets = buildEnv {
+      name = "runit-targets";
+      paths = [
+        # Startup
+        (writeScriptDir "1" ''
+          #!/bin/sh
+          echo 'Hello world!'
+          touch /etc/runit/stopit
+          chmod 0 /etc/runit/stopit
+        '')
+
+        # Run-time
+        (writeScriptDir "2" ''
+          #!/bin/sh
+          echo "Entering run-time"
+
+          cat /proc/uptime
+          echo "Running services in ${service_targets}..."
+          exec ${pkgsLinux.runit}/bin/runsvdir -P ${service_targets}
+        '')
+
+        # Shutdown
+        (writeScriptDir "3" ''
+          #!/bin/sh
+          echo 'Ok, bye...'
+        '')
+      ];
+    };
+
+    service_targets = buildEnv {
+      name = "service-targets";
+      paths = [
+        (writeRunitForegroundService "acpid" ''
+          #!/bin/sh
+          exec ${pkgsLinux.busybox}/bin/acpid -f
+        '')
+
+        (writeRunitForegroundService "sshd" ''
+          #!/bin/sh
+          exec ${pkgsLinux.openssh}/bin/sshd -D -e -f ${sshdConfig}
+        '')
+
+        (writeRunitForegroundService "vpnkit-forwarder" ''
+          #!/bin/sh
+          exec ${pkgsLinux.go-vpnkit}/bin/vpnkit-forwarder
+        '')
+
+        (writeRunitForegroundService "postboot-instructions" ''
+          #!/bin/sh
+
+          sleep 1
+          inst() {
+            echo
+            echo -e '\033[91;47m'
+            cat ${file_instructions}
+            echo -e '\033[0m'
+          }
+
+          inst
+          while read x; do inst; done
+        '')
+      ];
+    };
+  in writeScript "vm-run-stage2" ''
+    #! ${pkgsLinux.bash}/bin/bash -eux
 
     export NIX_STORE=${storeDir}
     export NIX_BUILD_TOP=/tmp
@@ -123,22 +235,23 @@ let
     ${pkgsLinux.coreutils}/bin/ln -fs ${pkgsLinux.bash}/bin/sh /bin/sh
 
     # # Set up automatic kernel module loading.
-    export MODULE_DIR=${pkgsLinux.linux}/lib/modules/
-    ${pkgsLinux.coreutils}/bin/cat <<EOF > /run/modprobe
-    #! /bin/sh
-    export MODULE_DIR=$MODULE_DIR
-    exec ${pkgsLinux.kmod}/bin/modprobe "\$@"
-    EOF
+    ${pkgsLinux.coreutils}/bin/cat ${script_modprobe} > /run/modprobe
     ${pkgsLinux.coreutils}/bin/chmod 755 /run/modprobe
     echo /run/modprobe > /proc/sys/kernel/modprobe
-    ${pkgsLinux.kmod}/bin/modprobe virtio_net
+    # This never passed, before:
+    # + /nix/store/q1yvjh3ab1sb5sn510zla8nlr2r1iimp-kmod-24/bin/modprobe virtio_net
+    # modprobe: FATAL: Module virtio_net not found in directory /nix/store/yimxiswgxpps9j19kykddfibj02b4k05-linux-4.9.50/lib/modules//4.9.50-linuxkit
+    # /run/modprobe virtio_net
 
-    echo "root:x:0:0:System administrator:/root:${pkgsLinux.bash}/bin/bash" >> /etc/passwd
-    echo "sshd:x:1:65534:SSH privilege separation user:/var/empty:${pkgsLinux.shadow}/bin/nologin" >> /etc/passwd
-    echo "nixbld1:x:30001:30000:Nix build user 1:/var/empty:${pkgsLinux.shadow}/bin/nologin" >> /etc/passwd
-    echo "nixbld:x:30000:nixbld1" >> /etc/group
+    cat ${file_passwd} > /etc/passwd
+    cat ${file_group} > /etc/group
 
-    export PATH="${vmToolsLinux.initrdUtils}/bin:${pkgsLinux.nix}/bin"
+    mkdir -p /etc/ssh /root /var/db /var/empty
+    chown root:root /root
+    chmod 0700 /root
+
+    cat ${file_bashrc} > /root/.bashrc
+    . /root/.bashrc
 
     if [ -f /nix-path-registration ]; then
       cat /nix-path-registration | nix-store --load-db
@@ -146,65 +259,56 @@ let
     fi
 
     ln -s /dev/pts/ptmx /dev/ptmx
-    mkdir -p /etc/ssh /root/.ssh /var/db /var/empty
-
     ln -s /proc/self/fd/0 /dev/stdin
 
     ifconfig eth0 ${containerIp}
     route add default gw 192.168.65.1 eth0
     echo 'nameserver 192.168.65.1' > /etc/resolv.conf
 
-    ${pkgsLinux.openssh}/bin/ssh-keygen -A
-    echo -n > /root/.bashrc
-    echo "export PATH=$PATH" >> /root/.bashrc
-    echo "export NIX_SSL_CERT_FILE='${pkgsLinux.cacert}/etc/ssl/certs/ca-bundle.crt'" >> /root/.bashrc
-    cp ${authorizedKeys} /root/.ssh/authorized_keys
-    chmod 0644 /root/.ssh/authorized_keys
+    mkdir -p /mnt
+    mknod /dev/sr0 b 11 0
+    ${pkgsLinux.busybox}/bin/mount /dev/sr0 /mnt
+
+    if [ ! -f /mnt/config ]; then
+      echo "FAIL FAIL FAIL"
+      echo "You must pass an SSH key data file via via a CDROM (ie: -data on linuxkit)"
+      exit 1
+    fi
+
+    mkdir /extract-ssh-keys
+    (
+      rm -rf /root/.ssh
+      mkdir -p /root/.ssh
+      chmod 0700 /root/.ssh
+
+      cd /extract-ssh-keys
+      tar -xf /mnt/config
+      chmod 0600 ./*
+      chmod 0644 ./*.pub
+
+      mv client.pub /root/.ssh/authorized_keys
+      chmod 0600 /root/.ssh/authorized_keys
+      chown root:root /root/.ssh/authorized_keys
+      mv ssh_host_* /etc/ssh/
+    )
+    rm -rf /extract-ssh-keys
 
     mkdir -p /port
     mount -v -t 9p -o trans=virtio,dfltuid=1001,dfltgid=50,version=9p2000 port /port
 
     mkdir -p /etc/acpi/PWRF /etc/acpi/events
-    cat >/etc/acpi/PWRF/00000080 <<EOF
-    #!/bin/sh
-    poweroff
-    EOF
-    chmod +x "$pkgdir"/etc/acpi/PWRF/00000080
-    # ${pkgsLinux.acpid}/bin/acpid -d
-
-    ${pkgsLinux.openssh}/bin/sshd -e -f ${sshdConfig}
-
-    ${pkgsLinux.go-vpnkit}/bin/vpnkit-forwarder &
-    ${pkgsLinux.go-vpnkit}/bin/vpnkit-expose-port \
-      -i \
-      -host-ip 127.0.0.1 -host-port ${hostPort} \
-      -container-ip 192.168.65.2 -container-port 22 \
-      -no-local-ip &
+    cat ${script_poweroff} > /etc/acpi/PWRF/00000080
+    chmod +x /etc/acpi/PWRF/00000080
 
     mkdir -p /dev/input /etc/acpi/PWRF /var/log
     mknod /dev/input/event0 c 13 64
-    cat <<EOF > /etc/acpi/PWRF/00000080
-    #!/bin/sh
-    poweroff -f
-    EOF
+    cat ${script_poweroff_f} > /etc/acpi/PWRF/00000080
     chmod +x /etc/acpi/PWRF/00000080
 
-    echo
-    echo -e '\033[91;47m======================================================================'
-    echo 'Remote builder has started.'
-    echo
-    echo 'If this is a fresh VM you need to run the following on the host:'
-    echo '    sudo ~/.nixpkgs/linuxkit-builder/root-init.sh'
-    echo
-    echo 'Then you can use the following to build x86_64-linux derivations:'
-    echo '    sudo $(cat ~/.nixpkgs/linuxkit-builder/env) nix-build'
-    echo
-    echo 'Exit this VM by running:'
-    echo '    kill $(cat ~/.nixpkgs/linuxkit-builder/nix-state/hyperkit.pid)'
-    echo '======================================================================'
-    echo -en '\033[0m'
+    rm -rf /etc/runit
+    cp -r ${runit_targets} /etc/runit/
 
-    exec ${pkgsLinux.busybox}/bin/acpid -f
+    exec ${pkgsLinux.runit}/bin/runit
   '';
 
   img = "bzImage";
@@ -248,31 +352,49 @@ let
     ln -fs ${initrd}/initrd $DIR/nix-initrd.img
     echo -n "console=ttyS0 panic=1 command=${stage2Init} loglevel=7 debug" > $DIR/nix-cmdline
 
-    [ ! -f $DIR/key ] && {
-      cp '${privateKey}' $DIR/key
-      chmod 0400 $DIR/key
-    }
-    echo "root@localhost x86_64-linux $DIR/key $CPUS 1 $FEATURES" > $DIR/remote-systems
+    if [ ! -d $DIR/keys ]; then
+      mkdir -p $DIR/keys
+      (
+        cd $DIR/keys
+        ssh-keygen -C "Nix LinuxKit Builder, Client" -N "" -f client
+        ssh-keygen -C "Nix LinuxKit Builder, Server" -f ssh_host_ecdsa_key -N "" -t ecdsa
 
-    cat <<EOF > $DIR/root-init.sh
-    #!/usr/bin/env bash
-    if [[ \$EUID -ne 0 ]]; then
-      echo "This script should be executed as root." 1>&2
-      exit 1
+        tar -cf server-config.tar client.pub ssh_host_ecdsa_key.pub ssh_host_ecdsa_key
+
+        echo -n '[localhost]:${hostPort} ' > known_host
+        cat ssh_host_ecdsa_key.pub >> known_host
+      )
     fi
-    chown root $HOME/.nixpkgs/linuxkit-builder/key
-    ssh-keygen -R '[localhost]:${hostPort}'
-    ssh-keyscan -t ecdsa-sha2-nistp256 -p ${hostPort} localhost >> /var/root/.ssh/known_hosts
-    EOF
-    chmod +x $DIR/root-init.sh
 
-    ENV=$DIR/env
-    echo -n > $ENV
-    echo "NIX_REMOTE=" >> $ENV
-    echo "NIX_SSHOPTS=-p${hostPort}" >> $ENV
-    echo "NIX_BUILD_HOOK=build-remote.pl" >> $ENV
-    echo "NIX_REMOTE_SYSTEMS=$DIR/remote-systems" >> $ENV
-    echo "NIX_CURRENT_LOAD=$DIR/current-load" >> $ENV
+    cp ${./example.nix} $DIR/example.nix
+
+
+    cat <<-EOF > $DIR/finish-setup.sh
+      #!/bin/sh
+      cat <<EOI
+      1. Add the following to /etc/nix/machines:
+
+        nix-linuxkit x86_64-linux $DIR/keys/client $CPUS 1 $FEATURES
+
+      2. Add the following to /var/root/.ssh/config:
+
+        Host nix-linuxkit
+           HostName localhost
+           User root
+           Port 24083
+           IdentityFile $DIR/keys/client
+           StrictHostKeyChecking yes
+           UserKnownHostsFile $DIR/keys/known_host
+           IdentitiesOnly yes
+
+      3. Try it out!
+
+        nix-build $DIR/example.nix
+
+
+    EOF
+
+    chmod +x $DIR/finish-setup.sh
 
     PATH="${vpnkit}/bin:$PATH"
     exec ${linuxkit}/bin/linuxkit run \
@@ -281,6 +403,7 @@ let
       -networking vpnkit \
       -ip ${containerIp} \
       -disk $DIR/nix-disk,size=$SIZE \
+      -data $DIR/keys/server-config.tar \
       -cpus $CPUS \
       -mem $MEM \
       $DIR/nix
